Review/corrections 3,5,6

+++ ugc_2/src/api/v1/reviews.py
@@ -2,7 +2,10 @@ from datetime import datetime
 from http import HTTPStatus
 
 from core.config import logger
-from fastapi import APIRouter, Depends, Request
+from core.error import DocumentExistsException
+from core.logger import extra
+from fastapi import APIRouter, Depends, HTTPException, Request
+from fastapi_jwt_auth import AuthJWT
 from models.events import Review, ReviewPosted
 from services.reviews import (ReviewsService, get_events_service,
                               review_serializer)
@@ -17,13 +20,19 @@ router = APIRouter()
 async def add_review(
     request: Request,
     event: Review,
+    authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
-    extra = {"tag": "fast_api_app"}
     logger.info(f"request add review {request}", extra=extra)
-    new_review = await service.add_event(
-        ReviewPosted(**event.dict(), created_at=datetime.now())
-    )
+    authorize.fresh_jwt_required()
+    user_id = authorize.get_jwt_subject()
+    logger.info(f"user {user_id} is authorized", extra=extra)
+    try:
+        new_review = await service.add_event(
+            ReviewPosted(**event.dict(), created_at=datetime.now(), user_id=user_id)
+        )
+    except DocumentExistsException:
+        raise HTTPException(status_code=HTTPStatus.CONFLICT)
     review = review_serializer(new_review)
     return review
 
@@ -33,10 +42,18 @@ async def update_review(
     request: Request,
     review_id: str,
     event: Review,
+    authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
-    extra = {"tag": "fast_api_app"}
     logger.info(f"request update review {request}", extra=extra)
+    authorize.fresh_jwt_required()
+    user_id = authorize.get_jwt_subject()
+    logger.info(f"user {user_id} is authorized", extra=extra)
+    review = await service.find_one(review_id)
+    if not review:
+        return HTTPStatus.NOT_FOUND
+    if review.get("user_id") != user_id:
+        return HTTPStatus.FORBIDDEN
     updated_review = await service.update(review_id, event)
     review = review_serializer(updated_review)
     return review
@@ -46,9 +63,17 @@ async def update_review(
 async def delete_review(
     request: Request,
     review_id: str,
+    authorize: AuthJWT = Depends(),
     service: ReviewsService = Depends(get_events_service),
 ):
-    extra = {"tag": "fast_api_app"}
     logger.info(f"request delete review {request}", extra=extra)
-    result = await service.delete(review_id)
-    return HTTPStatus.NO_CONTENT if result else HTTPStatus.NOT_FOUND
+    authorize.fresh_jwt_required()
+    user_id = authorize.get_jwt_subject()
+    logger.info(f"user {user_id} is authorized", extra=extra)
+    review = await service.find_one(review_id)
+    if not review:
+        return HTTPStatus.NOT_FOUND
+    if review.get("user_id") != user_id:
+        return HTTPStatus.FORBIDDEN
+    await service.delete(review_id)
+    return HTTPStatus.NO_CONTENT
